% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPar.R
\name{createPar}
\alias{createPar}
\title{Function that mixes parameter with defaults}
\usage{
createPar(init = NULL, pars = NULL)
}
\arguments{
\item{init}{named list with initial values}

\item{pars}{named list with parameter values}
}
\value{
Complete set of parameters that can be used in \code{\link{psfmod}}
}
\description{
Function that mixes parameter with defaults
}
\details{
The init list can have the following elements: \cr
 \itemize{
 \item{"BA"}{ Biomass species A}
 \item{"BB"}{ Biomass species A}
 \item{"S"}{ Nutrient availability in absence of plants}
 \item{"DA"}{ Litter biomass species A}
 \item{"DB"}{ Litter biomass species B}
 \item{"L0"}{ Light supply}
 \item{"Sa"}{ Proportional abundance of soil community attributed to species A}
 \item{"Dca"}{ Proportional effciency of decomposer community compared to monoculture species A}
 \item{"Dcb"}{ Proportional effciency of decomposer community compared to monoculture species A}
  }
 The pars list contains the model parameters. Parameters marked with AB are a vector with two entries,
 whereas the first one corresponds to species A and the second to species B.
 The list can have the following elements: \cr
   \itemize{
 \item{"gLC"}{ Maximum growth rate AB under light limitation}
 \item{"kLC"}{ Light availability at which half the maximal growth rates of AB are reached}
 \item{"gNC"}{ Maximum growth rate AB under nutrient limitation}
 \item{"kNC"}{ Nitrogen availability at which half the maximal growth rates of AB are reached}
 \item{"mC"}{ Mortality rate AB}
 \item{"a"}{ Turnover rate of nutrient supply}
 \item{"qNC"}{ Nitrogen content of tissue AB}
 \item{"rho"}{ Soil bulk density}
 \item{"l_Root"}{ Rooting depth}
 \item{"QNC"}{ Nitrogen content of AB litter at which it decomposes at rate dC}
 \item{"alphaNC"}{ Light-litter feedback coefficient AB}
 \item{"dC"}{ Litter decomposition rate AB}
 \item{"cRate"}{ Parameter determining the characteristic timescale of light dynamics}
 \item{"gamma_LC"}{ Light interception coefficient AB}
 \item{"alpha_LC"}{ Light-litter feedback coefficient AB}
 \item{"alpha_a"}{ Influence of Sa on A}
 \item{"alpha_b"}{ Influence of Sa on B}
 \item{"beta_a"}{ Influence of Sb on A}
 \item{"beta_b"}{ Influence of Sb on B}
 \item{"sccr"}{Vector with two vaues indicating the speed with which the soil cummunities react to changes
 in the conditions. The first value represents the decomposers, the second the soil competitors. A value of 1 means
 immediate reaction, whereas values < 1 indicate delayed reaction.}
 \item{"hfa"}{Determines whether the efiiciency of the soil composers is based on litter quality (value 0) or
 "home field advantage" (value 1)}
  }
}

