% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psfmod.R
\name{feedbackModel}
\alias{feedbackModel}
\title{Main feedback model}
\usage{
feedbackModel(t, y, mu)
}
\arguments{
\item{t}{t}

\item{y}{initial values (see Details)}

\item{mu}{parameters}
}
\value{
matrix with results.
}
\description{
Main feedback model
}
\details{
For further explanation of the parameter or the model see \code{\link{psfmod}} or the package vignette.
}

