% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Idea_HomeFieldAdvantage.R
\docType{data}
\name{mycol}
\alias{mycol}
\title{Visualization of temporal dynamics
================================}
\format{An object of class \code{character} of length 6.}
\usage{
mycol
}
\description{
Visualization of temporal dynamics
================================
}
\keyword{datasets}

