% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psfmod.R
\name{psfmod}
\alias{psfmod}
\title{Convenient wrapper function to call the model}
\usage{
psfmod(pars, days, newVal = NULL, out = NULL)
}
\arguments{
\item{pars}{list with parameter values as created by \code{\link{createPar}}}

\item{days}{number of days to run the modle}

\item{newVal}{named list with new parameter values. This can e.g. be used to simulatie mowing by removing biomass.}

\item{out}{output of previous simulation. If provided function will restart the simulation}
}
\value{
matrix with the results. Can be diplayed using \code{\link{psfplot}}
}
\description{
Convenient wrapper function to call the model
}
\examples{
# First definition of a paramater set. I choose to only modify som
# of the starting values
pars <- createPar(init = list("S" = 20, "L0" = 60))

# Now the simulation can be run
out <- psfmod(pars = pars, days = 5000)

# And the results plotted
psfplot(out)
}
\references{
Eppinga, Maarten B., et al. "Litter feedbacks, evolutionary change and exotic plant invasion." Journal of Ecology 99.2 (2011): 503-514. \cr\cr Bever, James D. "Soil community feedback and the coexistence of competitors: conceptual frameworks and empirical tests." New Phytologist 157.3 (2003): 465-473.
}

