% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psfplot.R
\name{psfplot}
\alias{psfplot}
\title{Plotting function}
\usage{
psfplot(out, ...)
}
\arguments{
\item{out}{output of \code{\link{psfmod}}}

\item{...}{futher arguments passed to matplot}
}
\value{
plot
}
\description{
Plotting function
}

