% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPar.R
\name{updatePar}
\alias{updatePar}
\title{Function to update par}
\usage{
updatePar(out, par, newVal = NULL)
}
\arguments{
\item{out}{output of simulation}

\item{par}{parameter object as created with \code{\link{createPar}}}

\item{newVal}{optional named list of new parameter values. This can e.g. be used to simulatie mowing by removing biomass.}
}
\description{
Function to update par
}

